refactor: transform `PresentationStore` into `AppContext`
